lwm2m: fix bug: add afterClass "await (3 sec, closed all ports lwm2server and bootstrapServer)"

+++ common/transport/lwm2m/src/main/java/org/thingsboard/server/transport/lwm2m/bootstrap/store/LwM2MBootstrapConfigStoreTaskProvider.java
@@ -173,7 +173,7 @@ public class LwM2MBootstrapConfigStoreTaskProvider implements BootstrapTaskProvi
                 if (shId instanceof Long) {
                     shortId = ((Long) shId).intValue();
                 } else {
-                    shortId = ((Integer) shId).intValue();
+                    shortId = (int) shId;
                 }
                 serverInstances.put(shortId, instance.getId());
             });
